tests: move 2D gaussian creation to util
